1 add a new file for predicate column 2 remove uselesss code
