+++ be/src/vec/columns/column_vector.cpp
@@ -360,88 +360,6 @@ void ColumnVector<T>::get_extremes(Field& min, Field& max) const {
     max = NearestFieldType<T>(cur_max);
 }
 
-template <typename T>
-ColumnPtr ColumnVector<T>::filter_by_selector(const uint16_t* sel, size_t sel_size, ColumnPtr* ptr) {
-    if (ptr == nullptr) {
-        auto res = this->create();
-        if (sel_size == 0) {
-            return res;
-        }
- 
-        res->reserve(sel_size);
-        Container& res_data = res->get_data();
-        //todo(wb) vectorized here
-        for (size_t i = 0; i < sel_size; i++) {
-            res_data.push_back(data[sel[i]]);
-        }
-        return res;
-    } else {
-        if (sel_size == 0) {
-            return nullptr;
-        }
-        MutableColumnPtr ptr_res = (*std::move(*ptr)).assume_mutable();
-        //todo(wb) add DCHECK type
-        auto& cv_ref = reinterpret_cast<vectorized::ColumnVector<T>&>(*ptr_res);
-        Container& res_data = cv_ref.get_data();
-        for (size_t i = 0; i < sel_size; i++) {
-            res_data.push_back(data[sel[i]]);
-        }
-        return nullptr;
-    }
-}
- 
-template <typename T>
-ColumnPtr ColumnVector<T>::filter_date_by_selector(const uint16_t* sel, size_t sel_size, ColumnPtr* ptr) {
-    if (ptr == nullptr) {
-        auto res = this->create();
-        if (sel_size == 0) {
-            return res;
-        }
- 
-        const T* data_pos = data.data();
-        for (size_t i = 0; i < sel_size; i++) {
-            const T val = data_pos[sel[i]];
-            const char* val_ptr = reinterpret_cast<const char*>(&val);
-        
-            uint64_t value = 0;
-            value = *(unsigned char*)(val_ptr + 2);
-            value <<= 8;
-            value |= *(unsigned char*)(val_ptr + 1);
-            value <<= 8;
-            value |= *(unsigned char*)(val_ptr);
-            DateTimeValue date;
-            date.from_olap_date(value);
-        
-            res->insert_data(reinterpret_cast<char*>(&date), 0);
-        }
- 
-        return res;
-    } else {
-        if (sel_size == 0) {
-            return nullptr;
-        }
- 
-        const T* data_pos = data.data();
-        for (size_t i = 0; i < sel_size; i++) {
-            const T val = data_pos[sel[i]];
-            const char* val_ptr = reinterpret_cast<const char*>(&val);
-        
-            uint64_t value = 0;
-            value = *(unsigned char*)(val_ptr + 2);
-            value <<= 8;
-            value |= *(unsigned char*)(val_ptr + 1);
-            value <<= 8;
-            value |= *(unsigned char*)(val_ptr);
-            DateTimeValue date;
-            date.from_olap_date(value);
-        
-            (*std::move(*ptr)).assume_mutable()->insert_data(reinterpret_cast<char*>(&date), 0);
-        }
- 
-    }
-    return nullptr;
-}
-
 /// Explicit template instantiations - to avoid code bloat in headers.
 template class ColumnVector<UInt8>;
 template class ColumnVector<UInt16>;
